Refactor code for better redability
